ui: simplify the excise tax section's presentation, vertical radios
We don't need to present the tax options as an unordered list and then
present it again as radio buttons. We can provide the same info using
only the radio buttons. Also, they look better stacked vertically.

The introduction of the dagger for the infused product notes causes
that line to have inconsistent vertical spacing compared to the other
items, so we add some CSS to give a more consistent spacing.